+++ concierge-core/src/main/java/com/grookage/concierge/core/engine/processors/ActivateConfigProcessor.java
@@ -10,7 +10,6 @@ import com.grookage.concierge.models.config.ConfigState;
 import com.grookage.concierge.models.exception.ConciergeCoreErrorCode;
 import com.grookage.concierge.models.exception.ConciergeException;
 import com.grookage.concierge.repository.ConciergeRepository;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import lombok.experimental.SuperBuilder;