+++ concierge-core/src/main/java/com/grookage/concierge/core/engine/processors/AppendConfigProcessor.java
@@ -12,7 +12,6 @@ import com.grookage.concierge.models.exception.ConciergeCoreErrorCode;
 import com.grookage.concierge.models.exception.ConciergeException;
 import com.grookage.concierge.models.ingestion.UpdateConfigRequest;
 import com.grookage.concierge.repository.ConciergeRepository;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.SneakyThrows;
 import lombok.experimental.SuperBuilder;