+++ concierge-aerospike/src/main/java/com/grookage/concierge/aerospike/repository/AerospikeRepository.java
@@ -3,6 +3,7 @@ package com.grookage.concierge.aerospike.repository;
 import com.grookage.concierge.aerospike.client.AerospikeConfig;
 import com.grookage.concierge.aerospike.client.AerospikeManager;
 import com.grookage.concierge.aerospike.storage.AerospikeRecord;
+import com.grookage.concierge.models.SearchRequest;
 import com.grookage.concierge.models.config.ConfigDetails;
 import com.grookage.concierge.models.config.ConfigKey;
 import com.grookage.concierge.models.config.ConfigState;
@@ -68,28 +69,33 @@ public class AerospikeRepository implements ConciergeRepository {
     }
 
     @Override
-    public boolean createdRecordExists(String namespace, String configName) {
-        return aerospikeManager.exists(namespace, configName, ConfigState.CREATED.name());
+    public boolean createdRecordExists(ConfigKey configKey) {
+        return aerospikeManager.exists(configKey.getOrgId(), configKey.getNamespace(),
+                configKey.getTenantId(), configKey.getConfigName());
     }
 
     @Override
     public List<ConfigDetails> getStoredRecords() {
-        return aerospikeManager.getRecords(Set.of(), Set.of(), Set.of())
+        return aerospikeManager.getRecords(SearchRequest.builder().build())
                 .stream().map(this::toConfigDetails).toList();
     }
 
     @Override
-    public List<ConfigDetails> getStoredRecords(Set<String> namespaces, Set<String> configNames, Set<ConfigState> configStates) {
-        return aerospikeManager.getRecords(namespaces, configNames,
-                        configStates.stream().map(Enum::name).collect(Collectors.toSet()))
+    public List<ConfigDetails> getStoredRecords(SearchRequest searchRequest) {
+        return aerospikeManager.getRecords(searchRequest)
                 .stream().map(this::toConfigDetails).toList();
     }
 
     @Override
     public void rollOverAndUpdate(ConfigDetails configDetails) {
-        final var newRecords = aerospikeManager.getRecords(Set.of(configDetails.getConfigKey().getNamespace()),
-                        Set.of(configDetails.getConfigKey().getConfigName()),
-                        Set.of(ConfigState.ACTIVATED.name()))
+        final var searchRequest = SearchRequest.builder()
+                .orgs(Set.of(configDetails.getConfigKey().getOrgId()))
+                .namespaces(Set.of(configDetails.getConfigKey().getNamespace()))
+                .tenants(Set.of(configDetails.getConfigKey().getTenantId()))
+                .configNames(Set.of(configDetails.getConfigKey().getConfigName()))
+                .configStates(Set.of(ConfigState.ACTIVATED))
+                .build();
+        final var newRecords = aerospikeManager.getRecords(searchRequest)
                 .stream().peek(each -> each.setConfigState(ConfigState.ROLLED)).collect(Collectors.toList());
         newRecords.add(toStorageRecord(configDetails));
         aerospikeManager.bulkUpdate(newRecords);