+++ concierge-core/src/main/java/com/grookage/concierge/core/engine/processors/CreateConfigProcessor.java
@@ -35,7 +35,7 @@ public class CreateConfigProcessor extends ConciergeProcessor {
                 .orElseThrow((Supplier<Throwable>) () -> ConciergeException.error(ConciergeCoreErrorCode.VALUE_NOT_FOUND));
         final var recordExists = getRepositorySupplier()
                 .get()
-                .createdRecordExists(createConfigRequest.getNamespace(), createConfigRequest.getConfigName());
+                .createdRecordExists(createConfigRequest.getConfigKey());
         if (recordExists) {
             log.error("There are already stored configs present with namespace {} and configName {}. Please try updating them instead",
                     createConfigRequest.getNamespace(), createConfigRequest.getConfigName());