+++ concierge-core/src/test/java/com/grookage/concierge/core/engine/CreateConfigProcessorTest.java
@@ -11,6 +11,8 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
+import java.util.Optional;
+
 class CreateConfigProcessorTest extends AbstractProcessorTest {
 
     @Override
@@ -30,7 +32,7 @@ class CreateConfigProcessorTest extends AbstractProcessorTest {
                 ConfigurationRequest.class);
         conciergeContext.addContext(ConfigurationRequest.class.getSimpleName(), createConfigRequest);
         ContextUtils.addConfigUpdaterContext(conciergeContext, getConfigUpdater());
-        Mockito.when(getConciergeRepository().createdRecordExists(Mockito.any(), Mockito.any()))
+        Mockito.when(getConciergeRepository().createdRecordExists(Mockito.any()))
                 .thenReturn(true);
         final var processor = getConciergeProcessor();
         Assertions.assertThrows(ConciergeException.class, () -> processor.process(conciergeContext));
@@ -44,7 +46,7 @@ class CreateConfigProcessorTest extends AbstractProcessorTest {
                 ConfigurationRequest.class);
         conciergeContext.addContext(ConfigurationRequest.class.getSimpleName(), createConfigRequest);
         ContextUtils.addConfigUpdaterContext(conciergeContext, getConfigUpdater());
-        Mockito.when(getConciergeRepository().createdRecordExists(Mockito.any(), Mockito.any()))
+        Mockito.when(getConciergeRepository().createdRecordExists(Mockito.any()))
                 .thenReturn(false);
         final var processor = getConciergeProcessor();
         processor.process(conciergeContext);