+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.1-RC12]
+
+- Introduced orgId and tenantId to configKey
+
+
 ## [0.0.1-RC11]
 
 - Moved to [Korg](https://github.com/grookage/korg) from Leia Refresher.